+++ MoPubSDK/Internal/Common/MPAdConfiguration.h
@@ -55,6 +55,8 @@ extern NSString * const kAdTypeClear;
 extern NSString * const kAdTypeNative;
 extern NSString * const kAdTypeNativeVideo;
 
+extern NSString * const kClickthroughExperimentBrowserAgent;
+
 @interface MPAdConfiguration : NSObject
 
 @property (nonatomic, assign) MPAdType adType;